fix: overflow protection for BlakeRounds precompile limits count (#1817)

+++ tracer/arithmetization/src/main/java/net/consensys/linea/zktracer/module/blake2fmodexpdata/BlakeModexpData.java
@@ -63,7 +63,7 @@ public class BlakeModexpData implements OperationListModule<BlakeModexpDataOpera
   public void callBlake(final BlakeComponents blakeComponents, final int operationID) {
     operations.add(new BlakeModexpDataOperation(blakeComponents, operationID));
     blakeEffectiveCall.addPrecompileLimit(1);
-    blakeRounds.addPrecompileLimit(blakeComponents.r().toInt());
+    blakeRounds.addPrecompileLimit(blakeComponents.r());
     callWcpForIdCheck(operationID);
   }
